Changes to prepare for CRAN release
Fixed bug in compare.input.files.same.name when there were files with
the same name but different hashes.

Update roxygen comments

Created a testscripts directory to hold the scripts used by testthat so
it is easy to create new json when we want to test that it works with a
new release of rdtLite.

Updated copyright